requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.3
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.15.0=pyhd3eb1b0_0
aiohttp=3.8.1=py37h7f8727e_0
aiosignal=1.2.0=pyhd3eb1b0_0
albumentations=1.1.0=pypi_0
antlr4-python3-runtime=4.8=pypi_0
argcomplete=1.12.3=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
async-timeout=4.0.1=pyhd3eb1b0_0
asynctest=0.13.0=py_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pypi_0
black=21.12b0=pypi_0
blas=1.0=mkl
bleach=4.1.0=pypi_0
blinker=1.4=py37h06a4308_0
blosc=1.21.0=h8c45485_0
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py37h27cfd23_1003
brunsli=0.1=h2531618_0
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h27cfd23_0
ca-certificates=2024.12.31=h06a4308_0
cached-property=1.5.2=pypi_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.12.7=py37h06a4308_0
cffi=1.14.6=py37h400218f_0
cfitsio=3.470=hf0d0db6_6
charls=2.2.0=h2531618_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.3=pyhd3eb1b0_0
cloudpickle=1.3.0=pypi_0
cryptography=3.4.8=py37hd23ed53_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd3eb1b0_0
cython=0.29.26=pypi_0
cytoolz=0.11.0=py37h7b6447c_0
dask-core=2021.10.0=pyhd3eb1b0_0
dataclasses=0.6=pypi_0
debugpy=1.5.1=pypi_0
decorator=4.4.2=pypi_0
defusedxml=0.7.1=pypi_0
descartes=1.1.0=pypi_0
dictor=0.1.7=pypi_0
distro=1.9.0=pypi_0
docker-pycreds=0.4.0=pypi_0
efficientnet-pytorch=0.7.0=pyhd8ed1ab_0
einops=0.4.1=pypi_0
elementpath=1.3.3=pypi_0
entrypoints=0.3=pypi_0
ephem=4.1.3=pypi_0
et-xmlfile=1.1.0=pypi_0
ffmpeg=4.3=hf484d3e_0
fire=0.4.0=pypi_0
flake8=4.0.1=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
freia=0.2=pypi_0
frozenlist=1.2.0=py37h7f8727e_0
fsspec=2021.10.1=pyhd3eb1b0_0
future=0.18.2=py37_1
fvcore=0.1.2.post20201122=pyhd8ed1ab_0
giflib=5.2.1=h7b6447c_0
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=1.33.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
grpcio=1.42.0=py37hce63b2e_0
gym=0.17.2=pypi_0
h5py=3.6.0=pypi_0
hydra-core=1.1.1=pypi_0
idna=3.3=pyhd3eb1b0_0
imagecodecs=2021.8.26=py37h4cda21f_0
imageio=2.9.0=pyhd3eb1b0_0
imageio-ffmpeg=0.4.5=pypi_0
imgaug=0.4.0=pypi_0
importlib-metadata=4.2.0=pypi_0
importlib-resources=5.4.0=pypi_0
iniconfig=1.1.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.6.0=pypi_0
ipython=7.30.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.5=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.0.3=pypi_0
joblib=1.1.0=pypi_0
jpeg=9d=h7f8727e_0
jsonschema=4.3.2=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.1.0=pypi_0
jupyter-console=6.4.0=pypi_0
jupyter-core=4.9.1=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.2=pypi_0
jxrlib=1.1=h7b6447c_2
kiwisolver=1.3.1=py37h2531618_0
krb5=1.19.2=hac12032_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
lerc=3.0=h295c915_0
libaec=1.0.4=he6710b0_1
libcurl=7.80.0=h0b77cf5_0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libnghttp2=1.46.0=hce63b2e_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.17.2=h4ff587b_1
libssh2=1.9.0=h1ba5d50_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libzopfli=1.0.3=he6710b0_0
lmdb=1.3.0=pypi_0
locket=0.2.1=py37h06a4308_1
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py37h06a4308_0
markupsafe=2.0.1=pypi_0
matplotlib-base=3.5.0=py37h3ed280b_0
matplotlib-inline=0.1.3=pypi_0
mccabe=0.6.1=pypi_0
mistune=0.8.4=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.3.0=py37h54f3939_0
mkl_random=1.1.1=py37h0573a6f_0
motmetrics=1.1.3=pypi_0
moviepy=1.0.3=pypi_0
multidict=5.1.0=py37h27cfd23_2
munkres=1.1.4=py_0
mypy-extensions=0.4.3=pypi_0
nbclient=0.5.9=pypi_0
nbconvert=6.3.0=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.4=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=2.2=pypi_0
ninja=1.10.2=py37hd09550d_3
notebook=6.4.6=pypi_0
numpy=1.18.4=pypi_0
nuscenes-devkit=1.1.0=pypi_0
oauthlib=3.1.1=pyhd3eb1b0_0
olefile=0.46=py37_0
omegaconf=2.1.1=pypi_0
opencv-python=4.2.0.32=pypi_0
opencv-python-headless=4.5.1.48=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openpyxl=3.1.3=pypi_0
openssl=1.1.1w=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
partd=1.2.0=pyhd3eb1b0_0
pathspec=0.9.0=pypi_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.0.1=py37he98fc37_0
pip=21.0.1=py37h06a4308_0
platformdirs=2.4.1=pypi_0
plotly=5.5.0=pypi_0
pluggy=1.0.0=pypi_0
portalocker=2.3.0=py37h06a4308_0
proglog=0.1.9=pypi_0
prometheus-client=0.12.0=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.24=pypi_0
protobuf=3.17.2=py37h295c915_0
psutil=5.9.0=pypi_0
ptyprocess=0.7.0=pypi_0
py=1.11.0=pypi_0
py-trees=0.8.3=pypi_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycocotools=2.0.3=pypi_0
pycodestyle=2.8.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydot=1.4.2=pypi_0
pyflakes=2.4.0=pypi_0
pygame=2.5.2=pypi_0
pyglet=1.5.0=pypi_0
pygments=2.10.0=pypi_0
pyjwt=2.1.0=py37h06a4308_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyquaternion=0.9.9=pypi_0
pyrsistent=0.18.0=pypi_0
pysocks=1.7.1=py37_1
pytest=6.2.5=pypi_0
python=3.7.11=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.7=2_cp37m
pytorch=1.10.1=py3.7_cuda11.3_cudnn8.2.0_0
pytorch-lightning=1.2.5=pyhd8ed1ab_1
pytorch-mutex=1.0=cuda
pytz=2021.3=pypi_0
pywavelets=1.1.1=py37h7b6447c_2
pyyaml=6.0=py37h7f8727e_1
pyzmq=22.3.0=pypi_0
qtconsole=5.2.2=pypi_0
qtpy=2.0.0=pypi_0
qudida=0.0.4=pypi_0
readline=8.1=h27cfd23_0
requests=2.26.0=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-image=0.18.1=py37ha9443f7_0
scikit-learn=1.0.2=pypi_0
scipy=1.5.2=py37h0b6359f_0
send2trash=1.8.0=pypi_0
sentry-sdk=1.5.8=pypi_0
setproctitle=1.2.2=pypi_0
setuptools=49.6.0=py37_0
shapely=1.7.1=pypi_0
shortuuid=1.0.8=pypi_0
simple-watchdog-timer=0.1.1=pypi_0
six=1.16.0=pyhd3eb1b0_0
smmap=5.0.0=pypi_0
snappy=1.1.8=he6710b0_0
sqlite=3.36.0=hc218d9a_0
stable-baselines3=0.8.0=pypi_0
tabulate=0.8.9=py37h06a4308_0
tenacity=8.0.1=pypi_0
tensorboard=2.4.0=pyhc547734_0
tensorboard-plugin-wit=1.6.0=py_0
termcolor=1.1.0=py37h06a4308_1
terminado=0.12.1=pypi_0
testpath=0.5.0=pypi_0
threadpoolctl=3.0.0=pypi_0
tifffile=2021.7.2=pyhd3eb1b0_2
timm=0.5.4=pypi_0
tk=8.6.11=h1ccaba5_0
toml=0.10.2=pypi_0
tomli=1.2.3=pypi_0
toolz=0.11.2=pyhd3eb1b0_0
torchmetrics=0.6.2=pyhd8ed1ab_0
torchsummary=1.5.1=pypi_0
torchvision=0.11.2=py37_cu113
tornado=6.1=pypi_0
tqdm=4.50.2=py_0
traitlets=5.1.1=pypi_0
transforms3d=0.4.1=pypi_0
typed-ast=1.5.1=pypi_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.7=pyhd3eb1b0_0
wandb=0.12.11=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=2.0.2=pyhd3eb1b0_0
wheel=0.37.0=pyhd3eb1b0_1
widgetsnbextension=3.5.2=pypi_0
xmlschema=1.0.18=pypi_0
xz=5.2.5=h7b6447c_0
yacs=0.1.6=pyhd3eb1b0_1
yaml=0.2.5=h7b6447c_0
yarl=1.6.3=py37h27cfd23_0
yaspin=2.1.0=pypi_0
zfp=0.5.5=h2531618_6
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
